fix: remove compilation warnings and remove prints

+++ lib/graph.hpp
@@ -6,7 +6,7 @@ using namespace std;
 class Graph {
 public:
   Graph(string filename);
-  int countCliquesSerial(int k);
+  int countCliquesSerial(long unsigned int k);
   int countCliquesParalelo(int k, int t);
   int countCliquesParaleloBalanceado(int k);
   void print();
@@ -16,7 +16,7 @@ public:
 private:
   vector<vector<int>> adj_list;
   vector<int> vertices;
-  void add_edge(int v, int w);
+  void add_edge(long unsigned int v, long unsigned int w);
   int isNeighbor(int v, int w);
   int connectToAll(vector<int> clique, int v);
   int isInClique(vector<int> clique, int v);
